+++ node/src/Commands.ts
@@ -602,6 +602,16 @@ export function createSInter(keys: string[]): redis_request.Command {
 /**
  * @internal
  */
+export function createSUnionStore(
+    destination: string,
+    keys: string[],
+): redis_request.Command {
+    return createCommand(RequestType.SUnionStore, [destination].concat(keys));
+}
+
+/**
+ * @internal
+ */
 export function createSIsMember(
     key: string,
     member: string,
@@ -1411,6 +1421,13 @@ export function createPfAdd(
 ): redis_request.Command {
     const args = [key, ...elements];
     return createCommand(RequestType.PfAdd, args);
+}
+
+/**
+ * @internal
+ */
+export function createPfCount(keys: string[]): redis_request.Command {
+    return createCommand(RequestType.PfCount, keys);
 }
 
 /**